core: Move translate_select() to select.rs
